+++ doc/changelog.rst
@@ -1,6 +1,23 @@
 Changelog
 =========
 
+v0.2.0
+------
+*(unreleased)*
+
+- :class:`~plydata.verbs.distinct` now uses `pandas.unique` instead of
+  :func:`numpy.unique`.
+
+- Added function :func:`~plydata.utils.Q` for quote non-pythonic column
+  names in a dataframe.
+
+
+v0.1.1
+------
+*(2017-04-11)*
+
+Re-release of *v0.1.0*
+
 v0.1.0
 ------
 *(2017-04-11)*